Reorganise client tests and files
I found this confusing when first reading the code as it wasn't
clear there are two layers: a "client" and a cache wrapper, with
the latter being hidden in the __init__ file.

This rearranges the code to bring out the two layers and bring
together tests for all the clients so it's clear what we've got -
I also struggled to find the relevant test file.
