+++ tests/clients/antivirus/test_antivirus_client.py
