+++ tests/clients/encryption/test_encryption_client.py
