+++ tests/clients/redis/test_redis_client.py
