+++ tests/clients/redis/test_request_cache.py
